Stats functions that need dense chunks own 'densification'
